+++ resources/views/pages/evidence-browser-detail.blade.php
@@ -1,0 +1,8 @@
+<dl>
+    @foreach ($record as $name => $value)
+        @if ($value !== null)
+            <dt>{{ str($name)->replace('_', ' ')->ucfirst() }}</dt>
+            <dd>{{ $value }}</dd>
+        @endif
+    @endforeach
+</dl>